refactor: modernize WidgetSettingsStore and extract defaults
