+++ src/store/widget-defaults.ts
@@ -1,0 +1,419 @@
+import type { WidgetConfig } from '../types/widget-settings';
+
+export const LINEAR_MAP_SIZES: Record<
+  string,
+  { designWidth: number; designHeight: number }
+> = {
+  horizontal: { designWidth: 400, designHeight: 40 },
+  vertical: { designWidth: 40, designHeight: 400 },
+};
+
+export const INPUT_TRACE_SIZES: Record<
+  string,
+  { designWidth: number; designHeight: number }
+> = {
+  horizontal: { designWidth: 400, designHeight: 220 },
+  vertical: { designWidth: 400, designHeight: 110 },
+};
+
+export const DEFAULT_WIDGETS: WidgetConfig[] = [
+  {
+    id: 'speed',
+    label: 'Speed (Gear, Speed, RPM)',
+    description: 'Speedometer with gear and RPM indicator.',
+    enabled: true,
+    x: 400,
+    y: 100,
+    width: 312,
+    height: 90,
+    designWidth: 312,
+    designHeight: 90,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: 'F10',
+    customSettings: {
+      speed: {
+        focusMode: 'speed',
+        rpmColorTheme: 'custom',
+        rpmColorLow: '#22c55e',
+        rpmColorMid: '#eab308',
+        rpmColorHigh: '#ef4444',
+        rpmColorLimit: '#ff4d00',
+        showPitPanel: true,
+        showRpmBar: true,
+        showTemps: true,
+        pitSpeedLimitOverride: null,
+      },
+    },
+  },
+  {
+    id: 'input-trace',
+    label: 'Input Trace (Throttle, Brake, Clutch)',
+    description: 'Live throttle, brake, and clutch inputs.',
+    enabled: false,
+    x: 400,
+    y: 300,
+    width: 400,
+    height: 220,
+    designWidth: 400,
+    designHeight: 220,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: 'F11',
+    customSettings: {
+      'input-trace': {
+        showThrottle: true,
+        showBrake: true,
+        showClutch: true,
+        throttleColor: '#00ff00',
+        brakeColor: '#ff3333',
+        clutchColor: '#3399ff',
+        barMode: 'horizontal',
+      },
+    },
+  },
+  {
+    id: 'proximity-radar',
+    label: 'Proximity Radar',
+    description: 'Visual radar for nearby traffic.',
+    enabled: false,
+    x: 600,
+    y: 300,
+    width: 200,
+    height: 300,
+    designWidth: 200,
+    designHeight: 300,
+    backgroundColor: 'transparent',
+    backgroundColorEdge: 'transparent',
+    hotkey: 'F6',
+    customSettings: {
+      'proximity-radar': {
+        visibilityMode: 'proximity',
+        proximityThreshold: 5,
+        hideDelay: 2,
+      },
+    },
+  },
+  {
+    id: 'radar-bar',
+    label: 'Radar Bar',
+    description: 'Full-width side proximity indicators.',
+    enabled: false,
+    x: 200,
+    y: 300,
+    width: 800,
+    height: 380,
+    designWidth: 800,
+    designHeight: 380,
+    backgroundColor: 'transparent',
+    backgroundColorEdge: 'transparent',
+    hotkey: 'F7',
+    customSettings: {
+      'radar-bar': {
+        visibilityMode: 'proximity',
+        proximityThreshold: 3,
+        hideDelay: 2,
+        barDisplayMode: 'both',
+      },
+    },
+  },
+  {
+    id: 'standings',
+    label: 'Standings',
+    description: 'Live session standings and intervals.',
+    enabled: false,
+    x: 50,
+    y: 50,
+    width: 640,
+    height: 450,
+    designWidth: 640,
+    designHeight: 450,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: 'F3',
+    customSettings: {
+      standings: {
+        enableClassCycling: false,
+        classCyclingToggleHotkey: '',
+        classPrevHotkey: '',
+        classNextHotkey: '',
+        showPosChange: true,
+        showColumnHeaders: true,
+        showSessionHeader: true,
+        showWeather: true,
+        showSOF: true,
+        showTotalDrivers: true,
+        showBrand: true,
+        showTire: true,
+        showIRatingBadge: true,
+        showClassBadge: true,
+        showIrChange: false,
+        showPitStops: true,
+        showLapsCompleted: false,
+        abbreviateNames: false,
+      },
+    },
+  },
+  {
+    id: 'relative',
+    label: 'Relative',
+    description: 'Gaps to cars ahead and behind you.',
+    enabled: false,
+    x: 50,
+    y: 300,
+    width: 420,
+    height: 400,
+    designWidth: 420,
+    designHeight: 400,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: 'F4',
+    customSettings: {
+      relative: {
+        showIRatingBadge: true,
+        showClassBadge: true,
+        showPitIndicator: true,
+        abbreviateNames: true,
+        showTrendIcon: true,
+      },
+    },
+  },
+  {
+    id: 'track-map',
+    label: 'Track Map',
+    description: 'Dynamic 2D map of the current circuit.',
+    enabled: false,
+    x: 800,
+    y: 50,
+    width: 400,
+    height: 400,
+    designWidth: 400,
+    designHeight: 400,
+    backgroundColor: 'transparent',
+    backgroundColorEdge: 'transparent',
+    hotkey: 'F5',
+    customSettings: {
+      'track-map': {
+        showLegend: true,
+        legendPosition: 'right',
+        showSectors: true,
+        showSectorTimes: true,
+        showSectorsOnMap: true,
+        rotationMode: 'fixed',
+        playerDotColor: '#ffffff',
+        showPlayerLabel: true,
+        leaderLabelMode: 'all',
+        trackStrokePx: 10,
+        trackBorderPx: 3,
+        sectorStrokePx: 6,
+        targetDotRadiusPx: 10,
+      },
+    },
+  },
+  {
+    id: 'linear-map',
+    label: 'Linear Map',
+    description: 'Progress bar of car track positions.',
+    enabled: false,
+    x: 50,
+    y: 820,
+    width: 400,
+    height: 40,
+    designWidth: 400,
+    designHeight: 40,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: '',
+    customSettings: {
+      'linear-map': {
+        orientation: 'horizontal',
+        playerDotColor: '#ffffff',
+        targetDotRadiusPx: 10,
+      },
+    },
+  },
+  {
+    id: 'flags',
+    label: 'LED Flags',
+    description: 'LED matrix display of track flags.',
+    enabled: false,
+    x: 760,
+    y: 0,
+    width: 232,
+    height: 232,
+    designWidth: 232,
+    designHeight: 232,
+    backgroundColor: 'transparent',
+    backgroundColorEdge: 'transparent',
+    hotkey: '',
+    customSettings: {
+      flags: { alwaysShow: true, holdDuration: 3 },
+    },
+  },
+  {
+    id: 'flat-flags',
+    label: 'Flat Flags',
+    description: 'Banner-style list of active track flags.',
+    enabled: false,
+    x: 760,
+    y: 250,
+    width: 280,
+    height: 160,
+    designWidth: 280,
+    designHeight: 160,
+    backgroundColor: 'transparent',
+    backgroundColorEdge: 'transparent',
+    hotkey: '',
+    customSettings: {
+      'flat-flags': { alwaysShow: true, holdDuration: 3 },
+    },
+  },
+  {
+    id: 'chassis',
+    label: 'Chassis (Tires & Suspension)',
+    description: 'Tire pressures and brake temperatures.',
+    enabled: false,
+    x: 100,
+    y: 100,
+    width: 460,
+    height: 320,
+    designWidth: 460,
+    designHeight: 320,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: '',
+  },
+  {
+    id: 'example',
+    label: 'Telemetry Debug',
+    description: 'Raw telemetry data debugger.',
+    enabled: false,
+    x: 100,
+    y: 100,
+    width: 400,
+    height: 700,
+    designWidth: 400,
+    designHeight: 700,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: 'F8',
+  },
+  {
+    id: 'lap-delta',
+    label: 'Lap Delta',
+    description: 'Live delta against your best lap time.',
+    enabled: false,
+    x: 400,
+    y: 200,
+    width: 220,
+    height: 180,
+    designWidth: 220,
+    designHeight: 180,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: '',
+    customSettings: {
+      'lap-delta': {
+        layout: 'vertical',
+        showSectorTimes: true,
+        reference: 'session_best',
+      },
+    },
+  },
+  {
+    id: 'lap-times',
+    label: 'Lap Times',
+    description: 'Detailed history of your lap times.',
+    enabled: false,
+    x: 400,
+    y: 300,
+    width: 220,
+    height: 104,
+    designWidth: 220,
+    designHeight: 104,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: '',
+    customSettings: {
+      'lap-times': {
+        showLastLap: true,
+        showBestLap: true,
+        showP1: true,
+        layout: 'vertical',
+      },
+    },
+  },
+  {
+    id: 'timer',
+    label: 'Session Timer',
+    description: 'Stint and total session timers.',
+    enabled: false,
+    x: 50,
+    y: 310,
+    width: 240,
+    height: 120,
+    designWidth: 240,
+    designHeight: 120,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: '',
+    customSettings: {
+      timer: {
+        showFlag: true,
+        showLaps: true,
+        showPosition: true,
+        showWallClock: true,
+        showSimTime: true,
+        showPcDate: false,
+        showSimDate: true,
+      },
+    },
+  },
+  {
+    id: 'weather',
+    label: 'Weather',
+    description: 'Track conditions and wind information.',
+    enabled: false,
+    x: 760,
+    y: 200,
+    width: 240,
+    height: 280,
+    designWidth: 240,
+    designHeight: 280,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: '',
+    customSettings: {
+      weather: {
+        showCompass: true,
+        showAirTemp: true,
+        showTrackTemp: true,
+        showWind: true,
+        showHumidity: true,
+        showForecast: true,
+      },
+    },
+  },
+  {
+    id: 'fuel',
+    label: 'Fuel Strategy',
+    description: 'Fuel level and consumption calculator.',
+    enabled: false,
+    x: 760,
+    y: 500,
+    width: 240,
+    height: 360,
+    designWidth: 240,
+    designHeight: 360,
+    backgroundColor: '#252525',
+    backgroundColorEdge: '#14141b',
+    hotkey: '',
+    customSettings: {
+      fuel: {
+        showChart: false,
+        pitWarningLaps: 3,
+        chartType: 'bar',
+      },
+    },
+  },
+];
