+++ src/store/widget-settings.store.ts
@@ -1,6 +1,10 @@
-import { makeAutoObservable, reaction, runInAction } from 'mobx';
+import { makeAutoObservable, runInAction } from 'mobx';
 import { invoke } from '@tauri-apps/api/core';
-import { computedStore } from './iracing';
+import {
+  DEFAULT_WIDGETS,
+  LINEAR_MAP_SIZES,
+  INPUT_TRACE_SIZES,
+} from './widget-defaults';
 
 import type {
   FlagDisplaySettings,
@@ -21,426 +25,10 @@ import type {
   WidgetCustomSettings,
 } from '../types/widget-settings';
 
-const LINEAR_MAP_SIZES: Record<
-  string,
-  { designWidth: number; designHeight: number }
-> = {
-  horizontal: { designWidth: 400, designHeight: 40 },
-  vertical: { designWidth: 40, designHeight: 400 },
-};
-
-const INPUT_TRACE_SIZES: Record<
-  string,
-  { designWidth: number; designHeight: number }
-> = {
-  horizontal: { designWidth: 400, designHeight: 220 },
-  vertical: { designWidth: 400, designHeight: 110 },
-};
-
-export const DEFAULT_WIDGETS: WidgetConfig[] = [
-  {
-    id: 'speed',
-    label: 'Speed (Gear, Speed, RPM)',
-    description: 'Speedometer with gear and RPM indicator.',
-    enabled: true,
-    x: 400,
-    y: 100,
-    width: 312,
-    height: 90,
-    designWidth: 312,
-    designHeight: 90,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: 'F10',
-    customSettings: {
-      speed: {
-        focusMode: 'speed',
-        rpmColorTheme: 'custom',
-        rpmColorLow: '#22c55e',
-        rpmColorMid: '#eab308',
-        rpmColorHigh: '#ef4444',
-        rpmColorLimit: '#ff4d00',
-        showPitPanel: true,
-        showRpmBar: true,
-        showTemps: true,
-        pitSpeedLimitOverride: null,
-      },
-    },
-  },
-  {
-    id: 'input-trace',
-    label: 'Input Trace (Throttle, Brake, Clutch)',
-    description: 'Live throttle, brake, and clutch inputs.',
-    enabled: false,
-    x: 400,
-    y: 300,
-    width: 400,
-    height: 220,
-    designWidth: 400,
-    designHeight: 220,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: 'F11',
-    customSettings: {
-      'input-trace': {
-        showThrottle: true,
-        showBrake: true,
-        showClutch: true,
-        throttleColor: '#00ff00',
-        brakeColor: '#ff3333',
-        clutchColor: '#3399ff',
-        barMode: 'horizontal',
-      },
-    },
-  },
-  {
-    id: 'proximity-radar',
-    label: 'Proximity Radar',
-    description: 'Visual radar for nearby traffic.',
-    enabled: false,
-    x: 600,
-    y: 300,
-    width: 200,
-    height: 300,
-    designWidth: 200,
-    designHeight: 300,
-    backgroundColor: 'transparent',
-    backgroundColorEdge: 'transparent',
-    hotkey: 'F6',
-    customSettings: {
-      'proximity-radar': {
-        visibilityMode: 'proximity',
-        proximityThreshold: 5,
-        hideDelay: 2,
-      },
-    },
-  },
-  {
-    id: 'radar-bar',
-    label: 'Radar Bar',
-    description: 'Full-width side proximity indicators.',
-    enabled: false,
-    x: 200,
-    y: 300,
-    width: 800,
-    height: 380,
-    designWidth: 800,
-    designHeight: 380,
-    backgroundColor: 'transparent',
-    backgroundColorEdge: 'transparent',
-    hotkey: 'F7',
-    customSettings: {
-      'radar-bar': {
-        visibilityMode: 'proximity',
-        proximityThreshold: 3,
-        hideDelay: 2,
-        barDisplayMode: 'both',
-      },
-    },
-  },
-  {
-    id: 'standings',
-    label: 'Standings',
-    description: 'Live session standings and intervals.',
-    enabled: false,
-    x: 50,
-    y: 50,
-    width: 640,
-    height: 450,
-    designWidth: 640,
-    designHeight: 450,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: 'F3',
-    customSettings: {
-      standings: {
-        enableClassCycling: false,
-        classCyclingToggleHotkey: '',
-        classPrevHotkey: '',
-        classNextHotkey: '',
-        showPosChange: true,
-        showColumnHeaders: true,
-        showSessionHeader: true,
-        showWeather: true,
-        showSOF: true,
-        showTotalDrivers: true,
-        showBrand: true,
-        showTire: true,
-        showIRatingBadge: true,
-        showClassBadge: true,
-        showIrChange: false,
-        showPitStops: true,
-        showLapsCompleted: false,
-        abbreviateNames: false,
-      },
-    },
-  },
-  {
-    id: 'relative',
-    label: 'Relative',
-    description: 'Gaps to cars ahead and behind you.',
-    enabled: false,
-    x: 50,
-    y: 300,
-    width: 420,
-    height: 400,
-    designWidth: 420,
-    designHeight: 400,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: 'F4',
-    customSettings: {
-      relative: {
-        showIRatingBadge: true,
-        showClassBadge: true,
-        showPitIndicator: true,
-        abbreviateNames: true,
-        showTrendIcon: true,
-      },
-    },
-  },
-  {
-    id: 'track-map',
-    label: 'Track Map',
-    description: 'Dynamic 2D map of the current circuit.',
-    enabled: false,
-    x: 800,
-    y: 50,
-    width: 400,
-    height: 400,
-    designWidth: 400,
-    designHeight: 400,
-    backgroundColor: 'transparent',
-    backgroundColorEdge: 'transparent',
-    hotkey: 'F5',
-    customSettings: {
-      'track-map': {
-        showLegend: true,
-        legendPosition: 'right',
-        showSectors: true,
-        showSectorTimes: true,
-        showSectorsOnMap: true,
-        rotationMode: 'fixed',
-        playerDotColor: '#ffffff',
-        showPlayerLabel: true,
-        leaderLabelMode: 'all',
-        trackStrokePx: 10,
-        trackBorderPx: 3,
-        sectorStrokePx: 6,
-        targetDotRadiusPx: 10,
-      },
-    },
-  },
-  {
-    id: 'linear-map',
-    label: 'Linear Map',
-    description: 'Progress bar of car track positions.',
-    enabled: false,
-    x: 50,
-    y: 820,
-    width: 400,
-    height: 40,
-    designWidth: 400,
-    designHeight: 40,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: '',
-    customSettings: {
-      'linear-map': {
-        orientation: 'horizontal',
-        playerDotColor: '#ffffff',
-        targetDotRadiusPx: 10,
-      },
-    },
-  },
-  {
-    id: 'flags',
-    label: 'LED Flags',
-    description: 'LED matrix display of track flags.',
-    enabled: false,
-    x: 760,
-    y: 0,
-    width: 232,
-    height: 232,
-    designWidth: 232,
-    designHeight: 232,
-    backgroundColor: 'transparent',
-    backgroundColorEdge: 'transparent',
-    hotkey: '',
-    customSettings: {
-      flags: { alwaysShow: true, holdDuration: 3 },
-    },
-  },
-  {
-    id: 'flat-flags',
-    label: 'Flat Flags',
-    description: 'Banner-style list of active track flags.',
-    enabled: false,
-    x: 760,
-    y: 250,
-    width: 280,
-    height: 160,
-    designWidth: 280,
-    designHeight: 160,
-    backgroundColor: 'transparent',
-    backgroundColorEdge: 'transparent',
-    hotkey: '',
-    customSettings: {
-      'flat-flags': { alwaysShow: true, holdDuration: 3 },
-    },
-  },
-  {
-    id: 'chassis',
-    label: 'Chassis (Tires & Suspension)',
-    description: 'Tire pressures and brake temperatures.',
-    enabled: false,
-    x: 100,
-    y: 100,
-    width: 460,
-    height: 320,
-    designWidth: 460,
-    designHeight: 320,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: '',
-  },
-  {
-    id: 'example',
-    label: 'Telemetry Debug',
-    description: 'Raw telemetry data debugger.',
-    enabled: false,
-    x: 100,
-    y: 100,
-    width: 400,
-    height: 700,
-    designWidth: 400,
-    designHeight: 700,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: 'F8',
-  },
-  {
-    id: 'lap-delta',
-    label: 'Lap Delta',
-    description: 'Live delta against your best lap time.',
-    enabled: false,
-    x: 400,
-    y: 200,
-    width: 220,
-    height: 180,
-    designWidth: 220,
-    designHeight: 180,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: '',
-    customSettings: {
-      'lap-delta': {
-        layout: 'vertical',
-        showSectorTimes: true,
-        reference: 'session_best',
-      },
-    },
-  },
-  {
-    id: 'lap-times',
-    label: 'Lap Times',
-    description: 'Detailed history of your lap times.',
-    enabled: false,
-    x: 400,
-    y: 300,
-    width: 220,
-    height: 104,
-    designWidth: 220,
-    designHeight: 104,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: '',
-    customSettings: {
-      'lap-times': {
-        showLastLap: true,
-        showBestLap: true,
-        showP1: true,
-        layout: 'vertical',
-      },
-    },
-  },
-  {
-    id: 'timer',
-    label: 'Session Timer',
-    description: 'Stint and total session timers.',
-    enabled: false,
-    x: 50,
-    y: 310,
-    width: 240,
-    height: 120,
-    designWidth: 240,
-    designHeight: 120,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: '',
-    customSettings: {
-      timer: {
-        showFlag: true,
-        showLaps: true,
-        showPosition: true,
-        showWallClock: true,
-        showSimTime: true,
-        showPcDate: false,
-        showSimDate: false,
-      },
-    },
-  },
-  {
-    id: 'weather',
-    label: 'Weather',
-    description: 'Track conditions and wind information.',
-    enabled: false,
-    x: 760,
-    y: 200,
-    width: 240,
-    height: 280,
-    designWidth: 240,
-    designHeight: 280,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: '',
-    customSettings: {
-      weather: {
-        showCompass: true,
-        showAirTemp: true,
-        showTrackTemp: true,
-        showWind: true,
-        showHumidity: true,
-        showForecast: true,
-      },
-    },
-  },
-  {
-    id: 'fuel',
-    label: 'Fuel Strategy',
-    description: 'Fuel level and consumption calculator.',
-    enabled: false,
-    x: 760,
-    y: 500,
-    width: 240,
-    height: 360,
-    designWidth: 240,
-    designHeight: 360,
-    backgroundColor: '#252525',
-    backgroundColorEdge: '#14141b',
-    hotkey: '',
-    customSettings: {
-      fuel: {
-        showChart: false,
-        pitWarningLaps: 3,
-        chartType: 'bar',
-      },
-    },
-  },
-];
-
 class WidgetSettingsStore {
-  widgets: WidgetConfig[] = DEFAULT_WIDGETS;
+  widgets = new Map<string, WidgetConfig>(
+    DEFAULT_WIDGETS.map((w) => [w.id, { ...w }])
+  );
 
   standingsActiveClassIndex = 0;
 
@@ -452,41 +40,34 @@ class WidgetSettingsStore {
     makeAutoObservable(this, { autoSizedWidgets: false } as never, {
       autoBind: true,
     });
+  }
 
-    // Auto-clamp active class index when number of classes changes
-    reaction(
-      () => this.getStandingsClassCount(),
-      (total) => {
-        if (total > 0 && this.standingsActiveClassIndex >= total) {
-          runInAction(() => {
-            this.standingsActiveClassIndex = Math.max(0, total - 1);
-          });
-        }
-      }
-    );
+  get allWidgets(): WidgetConfig[] {
+    return Array.from(this.widgets.values());
   }
 
   setTrackMapForceStartPending(pending: boolean) {
     this.isTrackMapForceStartPending = pending;
   }
 
-  private getStandingsClassCount(): number {
-    const entries = computedStore.standings?.entries ?? [];
-    return new Set(entries.map((e) => e.carClassId)).size;
+  clampStandingsActiveClassIndex(totalClasses: number) {
+    if (totalClasses > 0 && this.standingsActiveClassIndex >= totalClasses) {
+      this.standingsActiveClassIndex = Math.max(0, totalClasses - 1);
+    }
   }
 
-  cycleStandingsPrev() {
-    const total = this.getStandingsClassCount();
-    if (total <= 1) return;
-    const clamped = Math.min(this.standingsActiveClassIndex, total - 1);
-    this.standingsActiveClassIndex = clamped === 0 ? total - 1 : clamped - 1;
+  cycleStandingsPrev(totalClasses: number) {
+    if (totalClasses <= 1) return;
+    const clamped = Math.min(this.standingsActiveClassIndex, totalClasses - 1);
+    this.standingsActiveClassIndex =
+      clamped === 0 ? totalClasses - 1 : clamped - 1;
   }
 
-  cycleStandingsNext() {
-    const total = this.getStandingsClassCount();
-    if (total <= 1) return;
-    const clamped = Math.min(this.standingsActiveClassIndex, total - 1);
-    this.standingsActiveClassIndex = clamped === total - 1 ? 0 : clamped + 1;
+  cycleStandingsNext(totalClasses: number) {
+    if (totalClasses <= 1) return;
+    const clamped = Math.min(this.standingsActiveClassIndex, totalClasses - 1);
+    this.standingsActiveClassIndex =
+      clamped === totalClasses - 1 ? 0 : clamped + 1;
   }
 
   toggleStandingsClassCycling() {
@@ -500,42 +81,48 @@ class WidgetSettingsStore {
   }
 
   setWidgets(widgets: WidgetConfig[]) {
-    // Merge incoming widgets with defaults to ensure new fields/widgets are present
-    this.widgets = DEFAULT_WIDGETS.map((def) => {
-      const s = widgets.find((w) => w.id === def.id);
-      if (!s) return def;
+    runInAction(() => {
+      DEFAULT_WIDGETS.forEach((def) => {
+        const s = widgets.find((w) => w.id === def.id);
+        if (!s) {
+          this.widgets.set(def.id, { ...def });
+          return;
+        }
 
-      const defCS = def.customSettings ?? {};
-      const sCS = s.customSettings ?? {};
+        const defCS = def.customSettings ?? {};
+        const sCS = s.customSettings ?? {};
 
-      const mergedCustomSettings: WidgetCustomSettings = { ...defCS };
-      for (const key of Object.keys(sCS) as Array<keyof WidgetCustomSettings>) {
-        mergedCustomSettings[key] = {
-          ...defCS[key],
-          ...sCS[key],
-        } as never;
-      }
+        const mergedCustomSettings: WidgetCustomSettings = { ...defCS };
+        for (const key of Object.keys(sCS) as Array<
+          keyof WidgetCustomSettings
+        >) {
+          mergedCustomSettings[key] = {
+            ...defCS[key],
+            ...sCS[key],
+          } as never;
+        }
 
-      const autoSizedCurrent = this.autoSizedWidgets.has(def.id)
-        ? this.widgets.find((w) => w.id === def.id)
-        : null;
+        const autoSizedCurrent = this.autoSizedWidgets.has(def.id)
+          ? this.widgets.get(def.id)
+          : null;
 
-      return {
-        ...def,
-        ...s,
-        customSettings: mergedCustomSettings,
-        ...(autoSizedCurrent && {
-          width: autoSizedCurrent.width,
-          height: autoSizedCurrent.height,
-          designWidth: autoSizedCurrent.designWidth,
-          designHeight: autoSizedCurrent.designHeight,
-        }),
-      };
+        this.widgets.set(def.id, {
+          ...def,
+          ...s,
+          customSettings: mergedCustomSettings,
+          ...(autoSizedCurrent && {
+            width: autoSizedCurrent.width,
+            height: autoSizedCurrent.height,
+            designWidth: autoSizedCurrent.designWidth,
+            designHeight: autoSizedCurrent.designHeight,
+          }),
+        });
+      });
     });
   }
 
   getWidget(id: string): WidgetConfig | undefined {
-    return this.widgets.find((w) => w.id === id);
+    return this.widgets.get(id);
   }
 
   setWidgetEnabled(id: string, enabled: boolean) {
@@ -543,7 +130,7 @@ class WidgetSettingsStore {
   }
 
   updatePosition(id: string, x: number, y: number) {
-    const widget = this.widgets.find((w) => w.id === id);
+    const widget = this.getWidget(id);
     if (widget && (widget.x !== x || widget.y !== y)) {
       widget.x = x;
       widget.y = y;
@@ -551,7 +138,7 @@ class WidgetSettingsStore {
   }
 
   updateSize(id: string, width: number, height: number) {
-    const widget = this.widgets.find((w) => w.id === id);
+    const widget = this.getWidget(id);
     if (widget && (widget.width !== width || widget.height !== height)) {
       widget.width = width;
       widget.height = height;
@@ -560,7 +147,7 @@ class WidgetSettingsStore {
 
   updateAutoSize(id: string, width: number, height: number) {
     this.autoSizedWidgets.add(id);
-    const widget = this.widgets.find((w) => w.id === id);
+    const widget = this.getWidget(id);
     if (
       widget &&
       (widget.width !== width ||
@@ -580,14 +167,14 @@ class WidgetSettingsStore {
     field: K,
     value: WidgetConfig[K]
   ) {
-    const widget = this.widgets.find((w) => w.id === id);
+    const widget = this.getWidget(id);
     if (widget) {
       widget[field] = value;
     }
   }
 
   updateCustomSettings(id: string, settings: WidgetCustomSettings) {
-    const widget = this.widgets.find((w) => w.id === id);
+    const widget = this.getWidget(id);
     if (widget) {
       const prevSettings = widget.customSettings;
 
@@ -596,43 +183,7 @@ class WidgetSettingsStore {
         ...settings,
       };
 
-      if (
-        id === 'linear-map' &&
-        settings['linear-map'] &&
-        'orientation' in settings['linear-map']
-      ) {
-        const prevOrientation = prevSettings?.['linear-map']?.orientation;
-        const nextOrientation = settings['linear-map'].orientation;
-        if (prevOrientation !== nextOrientation) {
-          const size = LINEAR_MAP_SIZES[nextOrientation];
-          if (size) {
-            widget.designWidth = size.designWidth;
-            widget.designHeight = size.designHeight;
-          }
-          const prevW = widget.width;
-          widget.width = widget.height;
-          widget.height = prevW;
-        }
-      }
-
-      if (
-        id === 'input-trace' &&
-        settings['input-trace'] &&
-        'barMode' in settings['input-trace']
-      ) {
-        const prevBarMode =
-          prevSettings?.['input-trace']?.barMode ?? 'horizontal';
-        const nextBarMode = settings['input-trace'].barMode;
-        if (prevBarMode !== nextBarMode && nextBarMode !== 'hidden') {
-          const size = INPUT_TRACE_SIZES[nextBarMode];
-          if (size) {
-            widget.width = size.designWidth;
-            widget.height = size.designHeight;
-            widget.designWidth = size.designWidth;
-            widget.designHeight = size.designHeight;
-          }
-        }
-      }
+      this.handleWidgetSpecificResize(id, prevSettings, settings);
 
       if (id === 'fuel' && settings.fuel && 'pitWarningLaps' in settings.fuel) {
         void invoke('set_pit_warning_laps', {
@@ -642,199 +193,117 @@ class WidgetSettingsStore {
     }
   }
 
-  getSpeedSettings(): SpeedWidgetSettings {
-    const widget = this.getWidget('speed');
-    return (
-      widget?.customSettings?.speed ?? {
-        focusMode: 'speed',
-        rpmColorTheme: 'custom',
-        rpmColorLow: '#22c55e',
-        rpmColorMid: '#eab308',
-        rpmColorHigh: '#ef4444',
-        rpmColorLimit: '#ff4d00',
-        showPitPanel: true,
-        showRpmBar: true,
-        showTemps: true,
-        pitSpeedLimitOverride: null,
+  private handleWidgetSpecificResize(
+    id: string,
+    prevSettings: WidgetCustomSettings | undefined,
+    newSettings: WidgetCustomSettings
+  ) {
+    const widget = this.getWidget(id);
+    if (!widget) return;
+
+    if (id === 'linear-map' && newSettings['linear-map']?.orientation) {
+      const prevOrientation = prevSettings?.['linear-map']?.orientation;
+      const nextOrientation = newSettings['linear-map'].orientation;
+      if (prevOrientation !== nextOrientation) {
+        const size = LINEAR_MAP_SIZES[nextOrientation];
+        if (size) {
+          widget.designWidth = size.designWidth;
+          widget.designHeight = size.designHeight;
+        }
+        const prevW = widget.width;
+        widget.width = widget.height;
+        widget.height = prevW;
       }
-    );
+    }
+
+    if (id === 'input-trace' && newSettings['input-trace']?.barMode) {
+      const prevBarMode =
+        prevSettings?.['input-trace']?.barMode ?? 'horizontal';
+      const nextBarMode = newSettings['input-trace'].barMode;
+      if (prevBarMode !== nextBarMode && nextBarMode !== 'hidden') {
+        const size = INPUT_TRACE_SIZES[nextBarMode];
+        if (size) {
+          widget.width = size.designWidth;
+          widget.height = size.designHeight;
+          widget.designWidth = size.designWidth;
+          widget.designHeight = size.designHeight;
+        }
+      }
+    }
+  }
+
+  private getSettings<T>(id: string, key: keyof WidgetCustomSettings): T {
+    const widget = this.getWidget(id);
+    const def = DEFAULT_WIDGETS.find((w) => w.id === id);
+    const defaultSettings = def?.customSettings?.[key] as T;
+    return (widget?.customSettings?.[key] as T) ?? defaultSettings;
+  }
+
+  getSpeedSettings(): SpeedWidgetSettings {
+    return this.getSettings('speed', 'speed');
   }
 
   getInputTraceSettings(): InputTraceSettings {
-    const widget = this.getWidget('input-trace');
-    return (
-      widget?.customSettings?.['input-trace'] ?? {
-        showThrottle: true,
-        showBrake: true,
-        showClutch: true,
-        throttleColor: '#00ff00',
-        brakeColor: '#ff3333',
-        clutchColor: '#3399ff',
-        barMode: 'horizontal',
-      }
-    );
+    return this.getSettings('input-trace', 'input-trace');
   }
 
   getRadarSettings(id: 'proximity-radar' | 'radar-bar'): RadarSettings {
-    const widget = this.getWidget(id);
-    return (
-      widget?.customSettings?.[id] ?? {
-        visibilityMode: 'proximity',
-        proximityThreshold: id === 'proximity-radar' ? 5 : 3,
-        hideDelay: 2,
-      }
-    );
+    return this.getSettings(id, id);
   }
 
   getStandingsSettings(): StandingsWidgetSettings {
-    const widget = this.getWidget('standings');
-    return (
-      widget?.customSettings?.standings ?? {
-        enableClassCycling: false,
-        classCyclingToggleHotkey: '',
-        classPrevHotkey: '',
-        classNextHotkey: '',
-        showPosChange: true,
-        showColumnHeaders: true,
-        showSessionHeader: true,
-        showWeather: true,
-        showSOF: true,
-        showTotalDrivers: true,
-        showBrand: true,
-        showTire: true,
-        showIRatingBadge: true,
-        showClassBadge: true,
-        showIrChange: false,
-        showPitStops: true,
-        showLapsCompleted: false,
-        abbreviateNames: false,
-      }
-    );
+    return this.getSettings('standings', 'standings');
   }
 
   getRelativeSettings(): RelativeWidgetSettings {
-    const widget = this.getWidget('relative');
-    return (
-      widget?.customSettings?.relative ?? {
-        showIRatingBadge: true,
-        showClassBadge: true,
-        showPitIndicator: true,
-        showTrendIcon: true,
-        abbreviateNames: true,
-      }
-    );
+    return this.getSettings('relative', 'relative');
   }
 
   getTrackMapSettings(): TrackMapWidgetSettings {
-    const widget = this.getWidget('track-map');
-    const settings = widget?.customSettings?.['track-map'];
-    const showSectors = settings?.showSectors ?? true;
-
+    const settings = this.getSettings<TrackMapWidgetSettings>(
+      'track-map',
+      'track-map'
+    );
+    // Handle special default logic from original code
+    const showSectors = settings.showSectors ?? true;
     return {
-      showLegend: true,
-      legendPosition: 'right',
-      showSectors,
-      showSectorTimes: showSectors,
-      showSectorsOnMap: showSectors,
-      rotationMode: 'fixed',
-      playerDotColor: '#ffffff',
-      showPlayerLabel: true,
-      leaderLabelMode: 'all',
-      trackStrokePx: 10,
-      trackBorderPx: 3,
-      sectorStrokePx: 6,
-      targetDotRadiusPx: 10,
       ...settings,
+      showSectors,
+      showSectorTimes: settings.showSectorTimes ?? showSectors,
+      showSectorsOnMap: settings.showSectorsOnMap ?? showSectors,
     };
   }
 
   getLinearMapSettings(): LinearMapWidgetSettings {
-    const widget = this.getWidget('linear-map');
-    return (
-      widget?.customSettings?.['linear-map'] ?? {
-        orientation: 'horizontal',
-        playerDotColor: '#ffffff',
-        targetDotRadiusPx: 9,
-      }
-    );
+    return this.getSettings('linear-map', 'linear-map');
   }
 
   getWeatherSettings(): WeatherWidgetSettings {
-    const widget = this.getWidget('weather');
-    const defaults: WeatherWidgetSettings = {
-      showCompass: true,
-      showAirTemp: true,
-      showTrackTemp: true,
-      showWind: true,
-      showHumidity: true,
-      showForecast: true,
-    };
-    return { ...defaults, ...widget?.customSettings?.weather };
+    return this.getSettings('weather', 'weather');
   }
 
   getFuelSettings(): FuelWidgetSettings {
-    const widget = this.getWidget('fuel');
-    return (
-      widget?.customSettings?.fuel ?? {
-        showChart: false,
-        pitWarningLaps: 3,
-        chartType: 'bar',
-      }
-    );
+    return this.getSettings('fuel', 'fuel');
   }
 
   getLapTimesSettings(): LapTimesWidgetSettings {
-    const widget = this.getWidget('lap-times');
-    return (
-      widget?.customSettings?.['lap-times'] ?? {
-        showLastLap: true,
-        showBestLap: true,
-        showP1: true,
-        layout: 'vertical',
-      }
-    );
+    return this.getSettings('lap-times', 'lap-times');
   }
 
   getChassisSettings(): ChassisWidgetSettings {
-    const widget = this.getWidget('chassis');
-    return (
-      widget?.customSettings?.chassis ?? {
-        showInboard: true,
-      }
-    );
+    return this.getSettings('chassis', 'chassis');
   }
 
   getLapDeltaSettings(): LapDeltaWidgetSettings {
-    const widget = this.getWidget('lap-delta');
-    return (
-      widget?.customSettings?.['lap-delta'] ?? {
-        layout: 'vertical',
-        showSectorTimes: true,
-        reference: 'session_best',
-      }
-    );
+    return this.getSettings('lap-delta', 'lap-delta');
   }
 
   getTimerSettings(): TimerWidgetSettings {
-    const widget = this.getWidget('timer');
-    return (
-      widget?.customSettings?.timer ?? {
-        showFlag: true,
-        showLaps: true,
-        showPosition: true,
-        showWallClock: true,
-        showSimTime: true,
-        showPcDate: false,
-        showSimDate: true,
-      }
-    );
+    return this.getSettings('timer', 'timer');
   }
 
   getFlagDisplaySettings(id: 'flags' | 'flat-flags'): FlagDisplaySettings {
-    const widget = this.getWidget(id);
-    return (
-      widget?.customSettings?.[id] ?? { alwaysShow: true, holdDuration: 3 }
-    );
+    return this.getSettings(id, id);
   }
 }
